+++ src/namespaces/index.ts
@@ -3,6 +3,10 @@ export {
 } from './dec112';
 
 export {
+  EmergencySpecifics,
+} from './emergency';
+
+export {
   Mapper,
   Namespace,
   NamespaceSpecifics,